fix: resolve all code quality issues
- Fix macro_example compilation (use fn_sig instead of fn_inputs)
- Fix rediq-macros doc tests (use rust,ignore for examples)
- Fix all clippy warnings (type aliases, unnecessary casts, simplified expressions)
- Fix MutexGuard across await in priority_queue test
- Remove unused ErrorMiddleware warning with #[allow(dead_code)]
- Remove useless assert!(true) test in dependencies.rs

+++ rediq/src/server/worker.rs
@@ -485,7 +485,7 @@ impl Worker {
         ).await?;
 
         // Update TTL
-        let task_ttl = crate::config::get_task_ttl() as u64;
+        let task_ttl = crate::config::get_task_ttl();
         self.state.redis.expire(task_key, task_ttl).await?;
 
         // Update worker processed_total on successful completion
